pyproject.toml
[project]
name = "agendou-api"
version = "0.1.0"
description = ""
authors = [
    {name = "carloscabral656",email = "edu.carlos.cec@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg[binary] (>=3.2.0,<4.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
    "pyjwt[crypto] (>=2.8.0,<3.0.0)",
    "bcrypt (>=4.0.0,<6.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)"
]

[tool.poetry.scripts]
start = "agendou_api.main:run"

[tool.poetry]
packages = [{include = "agendou_api", from = "src"}]

[dependency-groups]
dev = [
    "ruff (>=0.15.9,<0.16.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
# FastAPI uses Depends() in defaults by design.
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
